+++ examples/rtic_bare1.rs
@@ -231,7 +231,7 @@ const APP: () = {
 //
 //    The programs tries to perform a carry operation and calls a function with "bl". Because it overflows it throws a panic. 
 //
-//    Commit your answer (bare1_3)
+//    Commit your answer (bare1_4)
 //
 // 5. Now we can remove the break point (click the `Remove All Breakpoints`),
 //    and instead uncomment the two breakpoint instructions (on lines 23 and 25).